to fix check_replace for
