+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: olardeux <olardeux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 15:37:52 by olardeux          #+#    #+#             */
-/*   Updated: 2024/11/08 23:54:39 by olardeux         ###   ########.fr       */
+/*   Updated: 2024/11/09 22:40:06 by olardeux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define MINISHELL_H
 
 # include "../libft/libft.h"
+# include "parsing.h"
 # include <errno.h>
 # include <fcntl.h>
 # include <readline/history.h>
@@ -41,33 +42,6 @@
 # endif
 
 # define PROMPT "minishell\033[0;34m$\033[0m "
-
-enum					e_token
-{
-	WORD,
-	INPUT,   // <
-	OUTPUT,  // >
-	APPEND,  // >>
-	HEREDOC, // <<
-	PIPE     // |
-};
-
-typedef struct s_token
-{
-	char				*token;
-	enum e_token		type;
-	struct s_token		*next;
-}						t_token;
-
-typedef struct s_parsing
-{
-	int					i;
-	int					args_count;
-	int					token_start;
-	int					token_count;
-	struct s_token		*tokens;
-	struct s_env		*env;
-}						t_parsing;
 
 typedef struct s_env
 {
@@ -97,35 +71,11 @@ typedef struct s_data
 
 char					*read_file(char *filename);
 
-t_cmd_list				*parsing(char **line, t_env *env);
-
-int						init_cmd_list(t_cmd_list **cmd_list, t_token *tokens);
-int						get_command(t_cmd_list *cmd, t_parsing *parsing,
-							t_token *start);
-int						detect_token(t_cmd_list *cmd, t_parsing *parsing,
-							t_token **start, int *j);
-
 t_env					*init_env(char **envp);
 char					*get_env_value(t_env *env, char *name);
 char					*get_env_exec(t_env *env, char *name);
 char					**env_to_array(t_env *env);
 t_env					*param_env(char *env);
-
-int						redirect_input(t_cmd_list *cmd, t_parsing *parsing,
-							t_token **start);
-int						redirect_output(t_cmd_list *cmd, t_parsing *parsing,
-							t_token **start);
-int						redirect_append(t_cmd_list *cmd, t_parsing *parsing,
-							t_token **start);
-
-t_token					*add_quote_token(t_token *tokens, char *line, int *i);
-int						quote_count_init(char *line, int *pos);
-
-t_token					*add_new_token(t_token *tokens);
-t_token					*add_token(t_parsing *parsing, char *line,
-							int token_len);
-t_token					*token_split(char *line);
-int						tokens_count(t_token *tokens);
 
 int						ft_isblank(char c);
 char					*ft_add_char(char *str, char c);
@@ -138,15 +88,10 @@ void					free_env(t_env *env);
 
 void					error_msg(char *msg, char *arg);
 
-char					*check_replace(char *line, t_env *env);
-
 void					signal_init(void);
 
 int						is_builtin(t_cmd_list *cmd);
 int						builtin(t_data *data);
-
-void					skip_quote(char *line, int *i);
-int						check_quote(char *line);
 
 int						name_match(char *name, char *arg);
